Reorganize test files per review feedback
- Move PtySession from tests/cli/pty.rs to distant-test-harness/src/pty.rs
  with portable-pty as a direct dependency of the harness crate
- Move tcp-echo-server smoke test to distant-test-harness/tests/
- Split ssh.rs into ssh.rs (shell only), connect.rs, and launch.rs
- Update all imports to use distant_test_harness::pty::PtySession
- Update TESTING.md to reflect new file locations
