+++ distant-test-harness/Cargo.toml
@@ -27,6 +27,7 @@ serde_json = "1"
 tokio = { version = "1", features = ["full"] }
 whoami = "2"
 which = "8"
+portable-pty = "0.9.0"
 rand = "0.8"
 rpassword = "7.4.0"
 ctrlc = "3"
