+++ distant-test-harness/src/lib.rs
@@ -3,6 +3,7 @@ pub mod exe;
 pub mod host;
 pub mod manager;
 pub mod process;
+pub mod pty;
 pub mod scripts;
 pub mod sshd;
 pub mod utils;
