+++ tests/cli/client/shell.rs
@@ -1,7 +1,7 @@
 //! Integration tests for the `distant shell` CLI subcommand.
 //!
 //! Uses `portable-pty` for cross-platform PTY session management via
-//! [`PtySession`](crate::cli::pty::PtySession). Tests shell execution,
+//! [`PtySession`](distant_test_harness::pty::PtySession). Tests shell execution,
 //! exit code forwarding, prediction modes, and alternate screen handling.
 
 use std::path::PathBuf;
@@ -34,7 +34,7 @@ async fn should_run_individual_command(#[case] backend: Backend) {
     let extra_args: &[&str] = &["--", "cmd", "/c", "echo", "hello"];
 
     let (bin, args) = shell_cmd_args(&ctx, extra_args);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.expect("hello");
     let exit_code = session.wait_for_exit();
@@ -54,7 +54,7 @@ async fn should_echo_input_through_pty(#[case] backend: Backend) {
         .expect("Failed to build pty-echo");
 
     let (bin, args) = shell_cmd_args(&ctx, &["--", &pty_echo_str]);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.send("abc");
     session.expect("abc");
@@ -73,7 +73,7 @@ async fn should_display_interactive_prompt(#[case] backend: Backend) {
         .expect("Failed to build pty-interactive");
 
     let (bin, args) = shell_cmd_args(&ctx, &["--", &pty_interactive_str]);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.expect("$ ");
 }
@@ -91,7 +91,7 @@ async fn should_exit_on_eof_signal(#[case] backend: Backend) {
         .expect("Failed to build pty-interactive");
 
     let (bin, args) = shell_cmd_args(&ctx, &["--", &pty_interactive_str]);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.expect("$ ");
 
@@ -135,7 +135,7 @@ async fn should_run_command_with_predict_off(#[case] backend: Backend) {
     ];
 
     let (bin, args) = shell_cmd_args(&ctx, extra_args);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.expect("predict-off-ok");
     let exit_code = session.wait_for_exit();
@@ -155,7 +155,7 @@ async fn should_handle_ctrl_c_interrupt(#[case] backend: Backend) {
         .expect("Failed to build pty-interactive");
 
     let (bin, args) = shell_cmd_args(&ctx, &["--", &pty_interactive_str]);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.expect("$ ");
     session.send("\x03");
@@ -177,7 +177,7 @@ async fn should_suppress_predicted_password_echo(#[case] backend: Backend) {
         .expect("Failed to build pty-password");
 
     let (bin, args) = shell_cmd_args(&ctx, &["--predict", "on", "--", &pty_password_str]);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.expect("Password: ");
     session.send_line("secret123");
@@ -206,7 +206,7 @@ async fn should_run_command_with_predict_on(#[case] backend: Backend) {
     ];
 
     let (bin, args) = shell_cmd_args(&ctx, extra_args);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.expect("predict-on-ok");
     let exit_code = session.wait_for_exit();
@@ -226,7 +226,7 @@ async fn should_not_echo_locally_with_predict_off(#[case] backend: Backend) {
         .expect("Failed to build pty-password");
 
     let (bin, args) = shell_cmd_args(&ctx, &["--predict", "off", "--", &pty_password_str]);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.expect("Password: ");
     session.send_line("secret123");
@@ -246,7 +246,7 @@ async fn should_echo_from_server_only_with_predict_off(#[case] backend: Backend)
         .expect("Failed to build pty-echo");
 
     let (bin, args) = shell_cmd_args(&ctx, &["--predict", "off", "--", &pty_echo_str]);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
     session.set_timeout(Duration::from_secs(60));
 
     session.expect("Connected to manager");
@@ -272,7 +272,7 @@ async fn should_echo_immediately_with_predict_on(#[case] backend: Backend) {
         .expect("Failed to build pty-echo");
 
     let (bin, args) = shell_cmd_args(&ctx, &["--predict", "on", "--", &pty_echo_str]);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.send("predict-immediate");
     session.expect("predict-immediate");
@@ -291,7 +291,7 @@ async fn should_correct_prediction_mismatch(#[case] backend: Backend) {
         .expect("Failed to build pty-password");
 
     let (bin, args) = shell_cmd_args(&ctx, &["--predict", "on", "--", &pty_password_str]);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.expect("Password: ");
     session.send_line("secret123");
@@ -328,7 +328,7 @@ async fn should_enter_alternate_screen(#[case] backend: Backend) {
     ];
 
     let (bin, args) = shell_cmd_args(&ctx, extra_args);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.expect("ALT_ENTRY_OK");
 }
@@ -363,7 +363,7 @@ async fn should_exit_alternate_screen(#[case] backend: Backend) {
     ];
 
     let (bin, args) = shell_cmd_args(&ctx, extra_args);
-    let mut session = crate::cli::pty::PtySession::spawn(&bin, &args);
+    let mut session = distant_test_harness::pty::PtySession::spawn(&bin, &args);
 
     session.expect("RESTORED");
 }
