+++ tests/cli/mod.rs
@@ -1,19 +1,16 @@
 //! CLI integration test module tree.
-//!
-//! Declares all submodules: `api` (JSON protocol tests), `client` (CLI subcommand tests),
-//! `generate`, `help`, `manager`, `pty`, `server`, and `tunnel`.
 
 mod api;
 mod client;
 mod config;
+mod connect;
 mod errors;
 mod format;
 mod generate;
 mod global_opts;
 mod help;
+mod launch;
 mod manager;
-mod pty;
 mod server;
 mod ssh;
-mod test_harness;
 mod tunnel;
